fix error copied file
